pyproject.toml
[project]
name = "p31s"
version = "0.1.0"
description = "Linux/macOS driver for POLONO P31S label printer"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
    {name = "P31S Printer Project"}
]
dependencies = [
    "bleak>=0.21.0",
    "pillow>=10.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pip-audit>=2.6.0",
]
barcodes = [
    "python-barcode>=0.15.0",
    "qrcode>=7.4.0",
]

[project.scripts]
p31s = "p31s.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "hardware: marks tests that require real printer hardware (deselect with '-m \"not hardware\"')",
    "barcodes: marks tests that require barcode dependencies (deselect with '-m \"not barcodes\"')",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.1",
]

[tool.ruff]
target-version = "py39"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["p31s"]
